Fix for Issue #32
patch serviceclient::get_auth_token_and_port() so that the home directory is specified externally to it.

Previously, it calculated the user home at the top of the function definition, but this caused issues when being run from non-Administrator accounts on Windows. The function would first be run as the non-Admin user.  The app detects that it doesn't have access to the system authtoken so it requests to elevate privileges and copy.  The issue arrises is that the destination path was re-calculated after running as an administrator so the target path of the copy was the administrator's home directory, not the initial user.  This patch sets the target home directory as a command line argument when running the copy_authtoken subcommand.